+++ x/intrarelayer/client/testutil/cli_test.go
@@ -1,0 +1,83 @@
+package testutil
+
+import (
+	"fmt"
+	"strings"
+	"testing"
+
+	tmcli "github.com/tendermint/tendermint/libs/cli"
+
+	"github.com/stretchr/testify/suite"
+
+	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
+
+	"github.com/tharsis/ethermint/testutil/network"
+	evmosnetwork "github.com/tharsis/evmos/testutil/network"
+	"github.com/tharsis/evmos/x/intrarelayer/client/cli"
+)
+
+type IntegrationTestSuite struct {
+	suite.Suite
+
+	cfg     network.Config
+	network *network.Network
+}
+
+func NewIntegrationTestSuite(cfg network.Config) *IntegrationTestSuite {
+	return &IntegrationTestSuite{cfg: cfg}
+}
+
+func TestIntegrationTestSuite(t *testing.T) {
+	suite.Run(t, new(IntegrationTestSuite))
+}
+
+func (s *IntegrationTestSuite) SetupSuite() {
+	s.T().Log("setting up integration test suite")
+
+	var err error
+	s.cfg = evmosnetwork.DefaultConfig()
+	s.cfg.NumValidators = 1
+
+	s.network, err = network.New(s.T(), s.T().TempDir(), s.cfg)
+	s.Require().NoError(err)
+	s.Require().NotNil(s.network)
+
+	_, err = s.network.WaitForHeight(1)
+	s.Require().NoError(err)
+}
+
+func (s *IntegrationTestSuite) TearDownSuite() {
+	s.T().Log("tearing down integration test suite")
+	s.network.Cleanup()
+}
+
+func (s *IntegrationTestSuite) TestCmdParams() {
+	val := s.network.Validators[0]
+
+	testCases := []struct {
+		name           string
+		args           []string
+		expectedOutput string
+	}{
+		{
+			"intrarelayer params",
+			[]string{
+				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
+			},
+			`{"params":{"enable_intrarelayer":true,"token_pair_voting_period":"172800s","enable_evm_hook":true}}`,
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+
+		s.Run(tc.name, func() {
+			cmd := cli.GetParamsCmd()
+			clientCtx := val.ClientCtx
+
+			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
+			s.Require().NoError(err)
+			s.Require().Equal(strings.TrimSpace(tc.expectedOutput), strings.TrimSpace(out.String()))
+		})
+	}
+}