+++ x/intrarelayer/types/proposal.go
@@ -40,10 +40,12 @@ func init() {
 	govtypes.RegisterProposalTypeCodec(&UpdateTokenPairERC20Proposal{}, "intrarelayer/UpdateTokenPairERC20Proposal")
 }
 
+// CreateDenomDescription generates a string with the coin description
 func CreateDenomDescription(address string) string {
 	return fmt.Sprintf("Cosmos coin token representation of %s", address)
 }
 
+// CreateDenom generates a string the module name plus the address to avoid conflicts with names staring with a number
 func CreateDenom(address string) string {
 	return fmt.Sprintf("%s/%s", ModuleName, address)
 }
@@ -175,12 +177,12 @@ func (etrp *ToggleTokenRelayProposal) ValidateBasic() error {
 }
 
 // NewUpdateTokenPairERC20Proposal returns new instance of UpdateTokenPairERC20Proposal
-func NewUpdateTokenPairERC20Proposal(title, description string, erc20Addr, newERC20Addr common.Address) govtypes.Content {
+func NewUpdateTokenPairERC20Proposal(title, description, erc20Addr, newERC20Addr string) govtypes.Content {
 	return &UpdateTokenPairERC20Proposal{
 		Title:           title,
 		Description:     description,
-		Erc20Address:    erc20Addr.String(),
-		NewErc20Address: newERC20Addr.String(),
+		Erc20Address:    erc20Addr,
+		NewErc20Address: newERC20Addr,
 	}
 }